rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "llama-index-core>=0.14.12",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "pymupdf>=1.23.0",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.21",
    "psycopg2-binary>=2.9.11",
    "pymongo[srv]>=4.16.0",
    "pyjwt>=2.10.1",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-llms-openrouter>=0.2.0",
    "langchain-google-genai>=4.2.1",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[dependency-groups]
dev = [
    "mongomock-motor>=0.0.36",
]
